pyproject.toml
[project]
name = "songbloom"
version = "0.1.0"
description = "Coherent Song Generation via Interleaved Autoregressive Sketching and Diffusion Refinement"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "cn2an==0.5.22",
    "descript-audio-codec==1.0.0",
    "einops==0.8.0",
    "g2p-en==2.1.0",
    "huggingface-hub==0.24.6",
    "jieba-fast==0.53",
    "lightning==2.2.1",
    "nltk==3.8.1",
    "num2words==0.5.13",
    "numpy<2",
    "omegaconf==2.2.0",
    "pyopenjtalk-plus>=0.4.1.post7",
    "pypinyin==0.51.0",
    "spacy==3.7.4",
    "torch==2.2.0",
    "torchaudio==2.2.0",
    "transformers==4.44.1",
    "vector-quantize-pytorch==1.14.8",
    "wordsegment==1.3.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
